refactor: centralize ticker normalization
